requirements.txt
easyocr
PyMuPDF
Pillow
numpy
schedule
torch
streamlit==1.45.0
pandas==2.2.3
plotly==5.21.0
sqlalchemy==2.0.40
